refactor: fix file name typo
