+++ src/components/inspection/InteractionButtons.tsx
